prototype/requirements.txt
faster-whisper
llama-cpp-python
sounddevice
numpy
scipy
huggingface_hub
requests
beautifulsoup4
onnxruntime
asyncio 
aiohttp